pyproject.toml
[project]
name = "gym-streamlit-dashboard"
version = "0.1.0"
description = "A Streamlit web application for uploading workout CSV data to Google BigQuery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "google-cloud-bigquery>=3.11.0",
    "google-auth>=2.22.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules"]
